+++ test/warpath/element/path_marker_test.exs
@@ -1,28 +1,32 @@
 defmodule Warpath.Element.PathMarkerTest do
   use ExUnit.Case, async: true
+
+  alias Warpath.Element
   alias Warpath.Element.PathMarker
 
   test "should generate index token for each item on list " do
     marked_paths =
-      {[2, 3], []}
-      |> PathMarker.stream(&(&2 ++ [&1]))
+      [2, 3]
+      |> Element.new([])
+      |> PathMarker.stream()
       |> Enum.to_list()
 
     assert marked_paths == [
-             {2, [{:index_access, 0}]},
-             {3, [{:index_access, 1}]}
+             Element.new(2, [{:index_access, 0}]),
+             Element.new(3, [{:index_access, 1}])
            ]
   end
 
   test "should generate property token for each key on map" do
     marked_paths =
-      {%{"name" => "Bumblebee", "group" => "Autobots"}, []}
-      |> PathMarker.stream(&(&2 ++ [&1]))
+      %{"name" => "Bumblebee", "group" => "Autobots"}
+      |> Element.new([])
+      |> PathMarker.stream()
       |> Enum.to_list()
 
     assert marked_paths == [
-             {"Autobots", [{:property, "group"}]},
-             {"Bumblebee", [{:property, "name"}]}
+             Element.new("Autobots", [{:property, "group"}]),
+             Element.new("Bumblebee", [{:property, "name"}])
            ]
   end
 end